Issue #41

* sync: fix windows.WAIT_TIMEOUT type.
* fifo: fix Close() method on Windows.
* gomod: update deps.

+++ sync/event_windows.go
@@ -57,7 +57,7 @@ func (e *WindowsEvent) WaitTimeout(timeout time.Duration) bool {
 	switch ev {
 	case windows.WAIT_OBJECT_0:
 		return true
-	case windows.WAIT_TIMEOUT:
+	case uint32(windows.WAIT_TIMEOUT):
 		return false
 	default:
 		if err != nil {
